Add doc string practice file
I'm in the middle of the Wizard's Adventure game from Land of Lisp, and
one bonus ability I want to add to the game is a command to list all
available commands for reference from the command line. Which requires
... learning how the built-in `documentation` function works!

Added a couple of stub functions with documentation strings, and wrote a
quick function to print out the doc strings for a given list of
functions.

Nice.