Add initial Wizard's Adventure game code
This is all of the game code from Land of Lisp through chapter 6. While
reading the book, you build a series of games which teach you the
fundamentals of ANSI Common Lisp, the first of which is a CLI
text-adventure game called Wizard's Adventure. This patch adds the code
for navigating through the rooms of the world, picking up items,
checking your inventory, looking around, and also the REPL that actually
runs the game.

The REPL I think blew my mind the most while writing this out. Some of
the Lisp-strategies for organizing and running your code I was familiar
with from working with Ramda.js, which is a Lisp-esque utility library
for JS. Treating and processing everything as though it were a list
wasn't new to me, but looking at the definition of a basic Lisp REPL
written in native Lisp is fairly wild:

```
;; this code totally works by the way:
(defun repl ()
  (loop (print (eval (read)))))

(repl)
```

Like ... what?! The combination of standard library functions plus the
syntax is just ... crazy. It's such a literal description of what's
actually going on: frist you read, then you evaluate, then you print,
and you're constantly looping all of that.

I like this language