Add docstring printing function
Here's my first adventure in writing non-codealong Lisp: a small
function which will print the docstrings for all of the allowed commands
in the game to remind you what they do and how they work.

Not very fancy, but I did learn about docstrings, and took that critical
step of writing code that I came up with myself.