Add cons cell practice
Experimenting with writing out my various thoughts along the way as I go
through the examples of a chapter. I think it makes it a little easier
to absorb the information. Kind of like annotating a book in college or
something.

Very cool to get deeper into lists. I'm looking forward to learning more
how Lisp handles tuples, which is a tool I've been very curious about
and have made some stabs at using in my everyday work, but I feel like I
have a shakey grasp of when they are the best tool for the job.